+++ components/my/profile/ProfileInfoTags/index.ts
@@ -1,0 +1,1 @@
+export { default } from './ProfileInfoTags';
